requirements.txt
# WealthOS Backend — Python dependencies
# Install: pip install -r requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# Data
pandas==2.2.3
numpy==2.1.3

# AI
groq==0.12.0

# Supabase
supabase==2.15.0
httpx==0.28.1

# Validation
pydantic==2.10.2